make contacts and calendar service optional
If enabled, incoming vcard and vcalendar items are stored in contact
service resp. calendar service. When disabled, such incoming items are
silently discarded. This follows the approach already taken for
vbookmark items.

The enabled/disabled check in CMake only looks at availability of the
.pc files. The .spec file then makes the .pc files available only when
desired. This is a bit error prone because something might pull them
in indirectly. However, that itself would be a bug and thus this
simple approach of extending CMake rules was used.

Only the mobile profile uses these services. Instead of checking for
the profile, a package or system bcond would be better. But that needs
to be fixed across multiple packages consistently and thus is not done
as part of this commit here.

Change-Id: I4eef2ab9cb20b9721becac5ea5989c74c183ab22
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>